+++ reikna/algorithms/pureparallel.py
@@ -115,6 +115,7 @@ class PureParallel(Computation):
 
         plan.kernel_call(
             template, args,
+            kernel_name="kernel_pure_parallel",
             global_size=(1,) if len(self._guiding_shape) == 0 else self._guiding_shape,
             render_kwds=dict(
                 idxs=idxs,