+++ reikna/core/transformation_modules.py
@@ -44,7 +44,7 @@ def flat_index_expr(param):
     names = index_cnames(param.annotation.type.shape)
 
     return " + ".join([
-        "(" + name + ")" + " * " + str(stride)
+        "(" + name + ")" + " * " + "(" + str(stride) + ")"
         for name, stride in zip(names, item_strides)]) + " + (" + str(item_offset) + ")"